+++ examples/appliance_device_snmp_v3_users.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2021] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -31,7 +31,7 @@ config = {
 
 options = {
     "type": "Users",
-    "userName": "user1",
+    "userName": "user1239",
     "securityLevel": "Authentication and privacy",
     "authenticationProtocol": "SHA512",
     "authenticationPassphrase": "authPass",
@@ -43,34 +43,38 @@ options = {
 config = try_load_from_file(config)
 
 oneview_client = OneViewClient(config)
+appliance_device_snmp_v3_users = oneview_client.appliance_device_snmp_v3_users
 
 # Add appliance device SNMP v3 users
-snmp_v3_user = oneview_client.appliance_device_snmp_v3_users.create(options)
-snmp_v3_user_uri = snmp_v3_user['uri']
+snmp_v3_user = appliance_device_snmp_v3_users.create(options)
+snmp_v3_user_uri = snmp_v3_user.data['uri']
 print("\n## Create appliance SNMP v3 user successfully!")
-pprint(snmp_v3_user)
+pprint(snmp_v3_user.data)
 
 # Lists the appliance SNMPv3 users
-snmp_v3_users_list = oneview_client.appliance_device_snmp_v3_users.get_all()
+snmp_v3_users_list = appliance_device_snmp_v3_users.get_all()
 print("\n## Got appliance SNMP v3 users successfully!")
 pprint(snmp_v3_users_list)
 
 # Get first element of the List
 snmp_v3_users = snmp_v3_users_list.pop()
 
+# Get by name
+print("\n## Find an SNMPv3 Users by username")
+snmp_v3_user = appliance_device_snmp_v3_users.get_by_name(snmp_v3_user.data['userName'])
+pprint(snmp_v3_user.data)
+
 # Get by URI
 print("Find an SNMP v3 user by URI")
-snmp_v3_user = oneview_client.appliance_device_snmp_v3_users.get(snmp_v3_user_uri)
-pprint(snmp_v3_user)
+snmp_v3_user = appliance_device_snmp_v3_users.get_by_uri(snmp_v3_user_uri)
+pprint(snmp_v3_user.data)
 
-# Change appliance device SNMP v3 users
-snmp_v3_user['authenticationPassphrase'] = "newAuthPass"
-snmp_v3_user['privacyPassphrase'] = "8765432187654321"
-snmp_v3_user = oneview_client.appliance_device_snmp_v3_users.update(snmp_v3_user)
-print("\n## Update appliance SNMP v3 user successfully!")
-pprint(snmp_v3_user)
+# hange appliance device SNMP v3 Users
+snmpv3_data = {"authenticationPassphrase": "newAuthPass", "privacyPassphrase": "8765432187654321"}
+snmp_v3_user = snmp_v3_user.update(snmpv3_data)
+print("\n## Update appliance SNMPv3 User successfully!")
+pprint(snmp_v3_user.data)
 
 # Delete Created Entry
-del_result = oneview_client.appliance_device_snmp_v3_users.delete(snmp_v3_user)
+snmp_v3_user.delete()
 print("\n## Delete appliance SNMP v3 user successfully!")
-pprint(del_result)