+++ CHANGELOG.md
@@ -5,9 +5,11 @@ Extends support to below features for REST API version 2400.
 #### Features supported
 - Appliance Configuration Timeconfig
 - Appliance SNMPv3 Trap Destinations
-- Appliance Time and Locale Configuration
+- Appliance SNMPv3  Users
 - Appliance SSH Access
+- Appliance Time and Locale Configuration
 - ID Pool IPv4 Range
+
 
 # 5.6.0
 #### Notes